requirements.txt
sentence-transformers
streamlit
redis
influxdb-client
ollama
python-dotenv
watchdog

